+++ pkgs/intl/lib/intl.dart
@@ -23,6 +23,7 @@ export 'src/intl/bidi_formatter.dart' show BidiFormatter;
 export 'src/intl/date_format.dart' show DateFormat;
 export 'src/intl/micro_money.dart' show MicroMoney;
 export 'src/intl/number_format.dart' show NumberFormat;
+export 'src/intl/number_parser_base.dart' show NumberParserBase;
 export 'src/intl/text_direction.dart' show TextDirection;
 
 /// The Intl class provides a common entry point for internationalization