+++ pkgs/intl/lib/src/intl/number_format.dart
@@ -4,6 +4,7 @@
 
 import 'dart:math';
 
+import 'package:intl/intl.dart';
 import 'package:intl/number_symbols.dart';
 import 'package:intl/number_symbols_data.dart';
 import 'package:intl/src/intl_helpers.dart' as helpers;
@@ -495,13 +496,30 @@ class NumberFormat {
 
   /// Parse the number represented by the string. If it's not
   /// parseable, throws a [FormatException].
-  num parse(String text) => NumberParser(this, text).value!;
+  num parse(String text) => parseWith(NumberParser.new, text);
+
+  /// Parse the number represented by the string using the parser created by the supplied parser generator. If it's not
+  /// parseable, throws a [FormatException].
+  R parseWith<R, P extends NumberParserBase<R>>(
+          P Function(NumberFormat, String) parserGenerator, String text) =>
+      parserGenerator(this, text).value!;
 
   /// Parse the number represented by the string. If it's not
   /// parsable, returns `null`.
   num? tryParse(String text) {
     try {
       return parse(text);
+    } on FormatException {
+      return null;
+    }
+  }
+
+  /// Parse the number represented by the string using the parser created by the supplied parser generator. If it's not
+  /// parsable, returns `null`.
+  R? tryParseWith<R, P extends NumberParserBase<R>>(
+      P Function(NumberFormat, String) parserGenerator, String text) {
+    try {
+      return parseWith(parserGenerator, text);
     } on FormatException {
       return null;
     }